+++ src/content_scripts/util.js
@@ -212,9 +212,13 @@ export async function appendMenu(options) {
       prevPageBtn.innerHTML = '<span class="oi" data-glyph="chevron-left" aria-hidden="true"></span>';
       prevPageBtn.className = (opts.url.previousPage) ? 'fmr-btn' : 'fmr-btn disabled';
       if (opts.url.previousPage) {
-        prevPageBtn.onclick = () => {
-          window.location.href = opts.url.previousPage;
-        };
+        if (typeof opts.url.previousPage === 'function') {
+          prevPageBtn.onclick = opts.url.previousPage;
+        } else {
+          prevPageBtn.onclick = () => {
+            window.location.href = opts.url.previousPage;
+          };
+        }
       }
     }
 
@@ -236,9 +240,13 @@ export async function appendMenu(options) {
       nextPageBtn.innerHTML = '<span class="oi" data-glyph="chevron-right"></span>';
       nextPageBtn.className = (opts.url.nextPage) ? 'fmr-btn' : 'fmr-btn disabled';
       if (opts.url.nextPage) {
-        nextPageBtn.onclick = () => {
-          window.location.href = opts.url.nextPage;
-        };
+        if (typeof opts.url.nextPage === 'function') {
+          nextPageBtn.onclick = opts.url.nextPage;
+        } else {
+          nextPageBtn.onclick = () => {
+            window.location.href = opts.url.nextPage;
+          };
+        }
       }
     }